refactor: rename passwordGenerator helper
